requirements.txt
hyperopt==0.2.7
lightgbm==3.3.3
matplotlib==3.5.2
mlens==0.2.3
notebook==6.5.1
numpy==1.21.6
pandas==1.3.5
plotly==5.10.0
pyarrow==8.0.0
scikit-learn==0.24.2
scipy==1.9.3
seaborn==0.12.1
shap==0.40.0
statsmodels==0.13.2
xgboost==1.5.2
sklearn-json

